refactor(github): split adapter write paths
