+++ internal/connector/github/adapter.go
@@ -9,7 +9,6 @@ import (
 	"sort"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/digitaldrywood/detent/internal/connector"
@@ -32,193 +31,6 @@ const (
 	defaultProjectItemStatusWriteParallelism  = 4
 	defaultProjectItemStatusWriteTimeout      = 2 * time.Minute
 )
-
-const projectItemsQuery = `
-query DetentGitHubProjectItems(
-  $projectId: ID!
-  $first: Int!
-  $after: String
-  $query: String
-) {
-  node(id: $projectId) {
-    ... on ProjectV2 {
-      items(first: $first, after: $after, query: $query) {
-        pageInfo { hasNextPage endCursor }
-        nodes {
-          id
-          content {
-            __typename
-            ... on Issue {
-              id
-              number
-              title
-              state
-              stateReason
-              url
-              repository { nameWithOwner }
-              labels(first: 20) { nodes { name } }
-            }
-          }
-          statusValue: fieldValueByName(name: "Status") {
-            ... on ProjectV2ItemFieldSingleSelectValue { name updatedAt }
-          }
-          priorityValue: fieldValueByName(name: "Priority") {
-            ... on ProjectV2ItemFieldSingleSelectValue { name }
-          }
-        }
-      }
-    }
-  }
-  rateLimit { limit used remaining cost resetAt }
-}`
-
-const observedStatusProjectItemsQuery = `
-query DetentGitHubObservedStatusProjectItems(
-  $projectId: ID!
-  $first: Int!
-  $after: String
-  $query: String
-) {
-  node(id: $projectId) {
-    ... on ProjectV2 {
-      items(first: $first, after: $after, query: $query) {
-        pageInfo { hasNextPage endCursor }
-        nodes {
-          id
-          content {
-            __typename
-            ... on Issue {
-              id
-              number
-              title
-              state
-              stateReason
-              url
-              repository { nameWithOwner }
-              closedByPullRequestsReferences(first: 5) { nodes { number url state repository { nameWithOwner } } }
-            }
-          }
-          statusValue: fieldValueByName(name: "Status") {
-            ... on ProjectV2ItemFieldSingleSelectValue { name updatedAt }
-          }
-          priorityValue: fieldValueByName(name: "Priority") {
-            ... on ProjectV2ItemFieldSingleSelectValue { name }
-          }
-        }
-      }
-    }
-  }
-  rateLimit { limit used remaining cost resetAt }
-}`
-
-const statusFieldQuery = `
-query DetentGitHubStatusField($projectId: ID!) {
-  node(id: $projectId) {
-    ... on ProjectV2 {
-      field(name: "Status") {
-        ... on ProjectV2SingleSelectField {
-          id
-          options { id name }
-        }
-      }
-    }
-  }
-  rateLimit { limit used remaining cost resetAt }
-}`
-
-const singleSelectFieldQuery = `
-query DetentGitHubProjectField($projectId: ID!, $fieldName: String!) {
-  node(id: $projectId) {
-    __typename
-    ... on ProjectV2 {
-      field(name: $fieldName) {
-        __typename
-        ... on ProjectV2Field {
-          id
-          dataType
-        }
-        ... on ProjectV2SingleSelectField {
-          id
-          options { id name color description }
-        }
-      }
-    }
-  }
-  rateLimit { limit used remaining cost resetAt }
-}`
-
-const projectItemForIssueQuery = `
-query DetentGitHubProjectItemForIssue($issueId: ID!, $projectItemsFirst: Int!, $after: String) {
-  node(id: $issueId) {
-    ... on Issue {
-      projectItems(first: $projectItemsFirst, after: $after) {
-        pageInfo { hasNextPage endCursor }
-        nodes {
-          id
-          project { id }
-          statusValue: fieldValueByName(name: "Status") {
-            ... on ProjectV2ItemFieldSingleSelectValue { name updatedAt }
-          }
-          priorityValue: fieldValueByName(name: "Priority") {
-            ... on ProjectV2ItemFieldSingleSelectValue { name }
-          }
-          fieldValues(first: 100) {
-            nodes {
-              __typename
-              ... on ProjectV2ItemFieldSingleSelectValue {
-                name
-                field { ... on ProjectV2FieldCommon { name } }
-              }
-              ... on ProjectV2ItemFieldTextValue {
-                text
-                field { ... on ProjectV2FieldCommon { name } }
-              }
-              ... on ProjectV2ItemFieldNumberValue {
-                number
-                field { ... on ProjectV2FieldCommon { name } }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-  rateLimit { limit used remaining cost resetAt }
-}`
-
-const updateSingleSelectFieldValueMutation = `
-mutation DetentGitHubUpdateSingleSelectField($projectId: ID!, $itemId: ID!, $fieldId: ID!, $optionId: String!) {
-  updateProjectV2ItemFieldValue(input: {
-    projectId: $projectId,
-    itemId: $itemId,
-    fieldId: $fieldId,
-    value: { singleSelectOptionId: $optionId }
-  }) {
-    projectV2Item { id }
-  }
-}`
-
-const updateTextFieldValueMutation = `
-mutation DetentGitHubUpdateTextField($projectId: ID!, $itemId: ID!, $fieldId: ID!, $text: String!) {
-  updateProjectV2ItemFieldValue(input: {
-    projectId: $projectId,
-    itemId: $itemId,
-    fieldId: $fieldId,
-    value: { text: $text }
-  }) {
-    projectV2Item { id }
-  }
-}`
-
-const deleteProjectItemMutation = `
-mutation DetentGitHubDeleteProjectItem($projectId: ID!, $itemId: ID!) {
-  deleteProjectV2Item(input: {
-    projectId: $projectId,
-    itemId: $itemId
-  }) {
-    deletedItemId
-  }
-}`
 
 var (
 	modelOverridePattern = regexp.MustCompile(`(?i)<!--\s*model:\s*(\S+?)\s*-->`)
@@ -530,348 +342,6 @@ type projectFieldValue struct {
 
 type projectField struct {
 	Name string `json:"name"`
-}
-
-func (c *Connector) CreateComment(ctx context.Context, issueID string, body string) error {
-	ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		return ErrCommentCreateFailed
-	}
-
-	var response struct {
-		NodeID string `json:"node_id"`
-	}
-	if err := c.client.REST(ctx, http.MethodPost, restIssueCommentsPath(ref), map[string]any{"body": body}, &response); err != nil {
-		return fmt.Errorf("create github comment: %w", err)
-	}
-	if strings.TrimSpace(response.NodeID) == "" {
-		return ErrCommentCreateFailed
-	}
-
-	return nil
-}
-
-func (c *Connector) CreateIssue(ctx context.Context, draft connector.IssueDraft) (connector.Issue, error) {
-	if !validPullRequestRepo(c.repository) {
-		return connector.Issue{}, ErrMissingRepository
-	}
-	title := strings.TrimSpace(draft.Title)
-	if title == "" {
-		return connector.Issue{}, ErrStatusUpdateFailed
-	}
-
-	body := strings.TrimSpace(draft.Body)
-	labels := normalizedIssueDraftLabels(draft.Labels)
-	payload := map[string]any{
-		"title": title,
-		"body":  body,
-	}
-	if len(labels) > 0 {
-		payload["labels"] = labels
-	}
-
-	var response restIssue
-	if err := c.client.REST(ctx, http.MethodPost, restRepositoryIssueCreatePath(c.repository), payload, &response); err != nil {
-		return connector.Issue{}, fmt.Errorf("create github issue: %w", err)
-	}
-	if strings.TrimSpace(response.NodeID) == "" || response.Number <= 0 {
-		return connector.Issue{}, ErrStatusUpdateFailed
-	}
-
-	ref := issueRef{Owner: c.repository.Owner, Name: c.repository.Name, Number: response.Number}
-	node := githubIssueNodeFromREST(ref, response)
-	c.cacheIssueRef(node)
-	status := c.labelStatusFromLabels(node.Labels)
-	if status == "" {
-		status = c.githubIssueStateToDetentState(node.State)
-	}
-	return c.buildLabelIssue(node, status), nil
-}
-
-func (c *Connector) CreatePullRequestComment(ctx context.Context, repository string, number int, body string) error {
-	owner, name, ok := splitRepositoryName(repository)
-	if !ok || number <= 0 {
-		return ErrCommentCreateFailed
-	}
-
-	var response struct {
-		NodeID string `json:"node_id"`
-	}
-	ref := issueRef{Owner: owner, Name: name, Number: number}
-	if err := c.client.REST(ctx, http.MethodPost, restIssueCommentsPath(ref), map[string]any{"body": body}, &response); err != nil {
-		return fmt.Errorf("create github pull request comment: %w", err)
-	}
-	if strings.TrimSpace(response.NodeID) == "" {
-		return ErrCommentCreateFailed
-	}
-
-	return nil
-}
-
-func normalizedIssueDraftLabels(labels []string) []string {
-	out := make([]string, 0, len(labels))
-	seen := map[string]struct{}{}
-	for _, label := range labels {
-		label = strings.TrimSpace(label)
-		if label == "" {
-			continue
-		}
-		key := strings.ToLower(label)
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		out = append(out, label)
-	}
-	return out
-}
-
-func (c *Connector) SetIssueField(ctx context.Context, issueID string, fieldID int, value string) error {
-	if fieldID <= 0 || strings.TrimSpace(value) == "" {
-		return ErrIssueFieldUpdateFailed
-	}
-
-	ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		return ErrIssueFieldUpdateFailed
-	}
-
-	var response []restIssueFieldValue
-	if err := c.client.REST(ctx, http.MethodPost, restIssueFieldValuesPath(ref), map[string]any{
-		"issue_field_values": []map[string]any{{
-			"field_id": fieldID,
-			"value":    strings.TrimSpace(value),
-		}},
-	}, &response); err != nil {
-		return fmt.Errorf("update github issue field: %w", err)
-	}
-	if len(response) == 0 {
-		return ErrIssueFieldUpdateFailed
-	}
-
-	return nil
-}
-
-func (c *Connector) ClearIssueField(ctx context.Context, issueID string, fieldID int) error {
-	if fieldID <= 0 || strings.TrimSpace(issueID) == "" {
-		return ErrIssueFieldUpdateFailed
-	}
-
-	ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		return ErrIssueFieldUpdateFailed
-	}
-
-	return c.deleteIssueFieldValue(ctx, ref, fieldID, ErrIssueFieldUpdateFailed)
-}
-
-func (c *Connector) CloseIssue(ctx context.Context, issueID string) error {
-	ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		return ErrIssueCloseFailed
-	}
-
-	var response restIssue
-	if err := c.client.REST(ctx, http.MethodPatch, restIssuePath(ref), map[string]any{
-		"state":        "closed",
-		"state_reason": "completed",
-	}, &response); err != nil {
-		return fmt.Errorf("close github issue: %w", err)
-	}
-	if strings.TrimSpace(response.NodeID) == "" || !strings.EqualFold(response.State, "closed") {
-		return ErrIssueCloseFailed
-	}
-
-	return nil
-}
-
-func (c *Connector) UpdateIssueState(ctx context.Context, issueID string, stateName string) error {
-	if c.usesLabelStatus() {
-		ref, ok, err := c.issueRefForID(ctx, strings.TrimSpace(issueID), graphQLQueryIssueLookup)
-		if err != nil {
-			return err
-		}
-		if !ok {
-			return ErrStatusUpdateFailed
-		}
-		issue, err := c.fetchRESTIssue(ctx, ref)
-		if err != nil {
-			return err
-		}
-		if strings.TrimSpace(issue.ID) == "" {
-			return ErrStatusUpdateFailed
-		}
-		currentStatus := c.labelStatusFromLabels(issue.Labels)
-		if currentStatus == "" {
-			currentStatus = c.githubIssueStateToDetentState(issue.State)
-		}
-		if c.terminalStatusUpdateBlocked(currentStatus, stateName) {
-			return c.stateUpdateBlockedError(issueID, currentStatus, stateName)
-		}
-		return c.updateIssueStatusLabel(ctx, ref, issue, stateName)
-	}
-	if c.usesIssueFieldStatus() {
-		ref, ok, err := c.issueRefForID(ctx, strings.TrimSpace(issueID), graphQLQueryIssueLookup)
-		if err != nil {
-			return err
-		}
-		if !ok {
-			return ErrStatusUpdateFailed
-		}
-		currentStatus, err := c.fetchIssueFieldStatus(ctx, ref)
-		if err != nil {
-			return err
-		}
-		if c.terminalStatusUpdateBlocked(currentStatus, stateName) {
-			return c.stateUpdateBlockedError(issueID, currentStatus, stateName)
-		}
-		githubState := c.detentToGitHubState(stateName)
-		return c.setIssueStatusField(ctx, ref, githubState)
-	}
-	if c.projectID == "" {
-		return ErrMissingProject
-	}
-
-	item, err := c.resolveProjectItem(ctx, strings.TrimSpace(issueID))
-	if err != nil {
-		return err
-	}
-	if c.terminalStatusUpdateBlocked(item.StatusName, stateName) {
-		return c.stateUpdateBlockedError(issueID, item.StatusName, stateName)
-	}
-
-	githubState := c.detentToGitHubState(stateName)
-	return c.setProjectItemStatus(ctx, item.ID, githubState)
-}
-
-func (c *Connector) RemoveIssueFromProject(ctx context.Context, issueID string) error {
-	issueID = strings.TrimSpace(issueID)
-	if issueID == "" {
-		return ErrProjectItemRemoveFailed
-	}
-	if c.usesLabelStatus() {
-		ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-		if err != nil {
-			return err
-		}
-		if !ok {
-			return ErrStatusUpdateFailed
-		}
-		return c.removeIssueStatusLabels(ctx, ref)
-	}
-	if c.usesIssueFieldStatus() {
-		ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-		if err != nil {
-			return err
-		}
-		if !ok {
-			return ErrStatusUpdateFailed
-		}
-		return c.clearIssueStatusField(ctx, ref)
-	}
-	if c.projectID == "" {
-		return ErrMissingProject
-	}
-
-	item, err := c.resolveProjectItem(ctx, issueID)
-	if err != nil {
-		return err
-	}
-	if err := c.deleteProjectItem(ctx, item.ID); err != nil {
-		return err
-	}
-	c.projectCache.ClearItemID(c.projectID, issueID)
-	return nil
-}
-
-func (c *Connector) setProjectItemStatus(ctx context.Context, itemID string, githubState string) error {
-	fieldID, optionID, err := c.resolveStatusOption(ctx, githubState)
-	if err != nil {
-		if errors.Is(err, ErrStatusOptionNotFound) {
-			c.statusCache.Clear(c.projectID)
-		}
-		return err
-	}
-
-	if err := c.updateStatusFieldValue(ctx, itemID, fieldID, optionID); err == nil {
-		return nil
-	}
-
-	c.statusCache.Clear(c.projectID)
-	fieldID, optionID, err = c.resolveStatusOption(ctx, githubState)
-	if err != nil {
-		return err
-	}
-	return c.updateStatusFieldValue(ctx, itemID, fieldID, optionID)
-}
-
-func (c *Connector) SetAssignee(ctx context.Context, issueID string, login string) error {
-	issueID = strings.TrimSpace(issueID)
-	login = strings.TrimSpace(login)
-	if issueID == "" || login == "" {
-		return ErrAssigneeNotFound
-	}
-	ref, ok, err := c.issueRefForID(ctx, issueID, graphQLQueryIssueLookup)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		return ErrAssigneeNotFound
-	}
-	currentAssignees, err := c.fetchIssueAssignees(ctx, ref)
-	if err != nil {
-		return err
-	}
-	removeLogins, alreadyAssigned := assigneeLoginReplacement(currentAssignees, login)
-	if !alreadyAssigned {
-		if err := c.addAssignee(ctx, ref, login); err != nil {
-			return err
-		}
-	}
-	if len(removeLogins) > 0 {
-		if err := c.removeAssignees(ctx, ref, removeLogins); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *Connector) SetField(ctx context.Context, issueID string, fieldName string, value string) error {
-	if c.usesIssueFieldStatus() {
-		return c.setIssueFieldValueByName(ctx, issueID, fieldName, value)
-	}
-	if c.projectID == "" {
-		return ErrMissingProject
-	}
-
-	item, err := c.resolveProjectItem(ctx, strings.TrimSpace(issueID))
-	if err != nil {
-		return err
-	}
-	field, err := c.fetchProjectField(ctx, fieldName)
-	if err != nil {
-		return err
-	}
-	if projectTextField(field) {
-		return c.updateProjectV2TextFieldValue(ctx, item.ID, field.ID, strings.TrimSpace(value), ErrProjectFieldUpdateFailed)
-	}
-	fieldID, optionID, err := c.resolveSingleSelectFieldOptionFromField(ctx, fieldName, value, field)
-	if err != nil {
-		return err
-	}
-	return c.updateProjectV2SingleSelectFieldValue(ctx, item.ID, fieldID, optionID, ErrProjectFieldUpdateFailed)
 }
 
 type issuePullRequestCandidate struct {
@@ -1780,314 +1250,6 @@ func (c *Connector) fetchPullRequestReviews(ctx context.Context, repo pullReques
 	return reviews, nil
 }
 
-func (c *Connector) fetchProjectItems(ctx context.Context, queryType string, query string, keepIssue func(connector.Issue) bool) ([]connector.Issue, error) {
-	return c.fetchProjectItemsLimit(ctx, queryType, query, keepIssue, 0)
-}
-
-func (c *Connector) fetchProjectBacklogIssues(ctx context.Context, limit int) ([]connector.Issue, error) {
-	issues, err := c.fetchExplicitProjectBacklogIssues(ctx, limit)
-	if err != nil {
-		return nil, err
-	}
-	if limit > 0 && len(issues) >= limit {
-		return issues, nil
-	}
-
-	blankLimit := 0
-	if limit > 0 {
-		blankLimit = limit - len(issues)
-	}
-	blankIssues, err := c.fetchBlankProjectBacklogIssues(ctx, blankLimit)
-	if err != nil {
-		return nil, err
-	}
-	return appendUniqueIssues(issues, blankIssues, limit), nil
-}
-
-func (c *Connector) fetchExplicitProjectBacklogIssues(ctx context.Context, limit int) ([]connector.Issue, error) {
-	backlogStates := []string{defaultProjectItemStatusState}
-	return c.fetchProjectItemsLimit(ctx, graphQLQueryCandidateIssues, c.projectStatusQuery(backlogStates), func(issue connector.Issue) bool {
-		return stateInList(issue.State, backlogStates)
-	}, limit)
-}
-
-func (c *Connector) fetchBlankProjectBacklogIssues(ctx context.Context, limit int) ([]connector.Issue, error) {
-	var after *string
-	issues := []connector.Issue{}
-	blankStatusItemIDs := []string{}
-
-	for {
-		var response struct {
-			Node *struct {
-				Items projectItemsConnection `json:"items"`
-			} `json:"node"`
-		}
-		if err := c.client.GraphQLWithType(ctx, graphQLQueryCandidateIssues, projectItemsQuery, map[string]any{
-			"projectId": c.projectID,
-			"first":     projectItemsPageSize,
-			"after":     after,
-			"query":     nil,
-		}, &response); err != nil {
-			return nil, fmt.Errorf("fetch github project items: %w", err)
-		}
-		if response.Node == nil {
-			return nil, ErrProjectNotFound
-		}
-
-		for _, item := range response.Node.Items.Nodes {
-			issue, ok, blankStatusItemID, err := c.normalizeProjectItem(item)
-			if err != nil {
-				return nil, err
-			}
-			if !ok || blankStatusItemID == "" {
-				continue
-			}
-			issues = append(issues, issue)
-			blankStatusItemIDs = append(blankStatusItemIDs, blankStatusItemID)
-			if limit > 0 && len(issues) >= limit {
-				c.defaultBlankProjectItemStatuses(ctx, blankStatusItemIDs)
-				return issues, nil
-			}
-		}
-
-		if !response.Node.Items.PageInfo.HasNextPage {
-			c.defaultBlankProjectItemStatuses(ctx, blankStatusItemIDs)
-			return issues, nil
-		}
-		cursor := strings.TrimSpace(response.Node.Items.PageInfo.EndCursor)
-		if cursor == "" {
-			return nil, ErrInvalidResponse
-		}
-		after = &cursor
-	}
-}
-
-func (c *Connector) fetchProjectItemsLimit(
-	ctx context.Context,
-	queryType string,
-	query string,
-	keepIssue func(connector.Issue) bool,
-	limit int,
-) ([]connector.Issue, error) {
-	return c.fetchProjectItemsWithLimit(ctx, projectItemsQueryForType(queryType), queryType, query, keepIssue, limit)
-}
-
-func (c *Connector) fetchProjectItemsWithPullRequestRefsLimit(
-	ctx context.Context,
-	queryType string,
-	query string,
-	keepIssue func(connector.Issue) bool,
-	limit int,
-) ([]connector.Issue, error) {
-	return c.fetchProjectItemsWithLimit(ctx, observedStatusProjectItemsQuery, queryType, query, keepIssue, limit)
-}
-
-func (c *Connector) fetchProjectItemsWithLimit(
-	ctx context.Context,
-	queryDocument string,
-	queryType string,
-	query string,
-	keepIssue func(connector.Issue) bool,
-	limit int,
-) ([]connector.Issue, error) {
-	var after *string
-	allIssues := []connector.Issue{}
-	blankStatusItemIDs := []string{}
-	var projectQuery *string
-	if query != "" {
-		projectQuery = &query
-	}
-
-	for {
-		var response struct {
-			Node *struct {
-				Items projectItemsConnection `json:"items"`
-			} `json:"node"`
-		}
-		if err := c.client.GraphQLWithType(ctx, queryType, queryDocument, map[string]any{
-			"projectId": c.projectID,
-			"first":     projectItemsPageSize,
-			"after":     after,
-			"query":     projectQuery,
-		}, &response); err != nil {
-			return nil, fmt.Errorf("fetch github project items: %w", err)
-		}
-		if response.Node == nil {
-			return nil, ErrProjectNotFound
-		}
-
-		for _, item := range response.Node.Items.Nodes {
-			issue, ok, blankStatusItemID, err := c.normalizeProjectItem(item)
-			if err != nil {
-				return nil, err
-			}
-			if !ok {
-				continue
-			}
-			if blankStatusItemID != "" {
-				blankStatusItemIDs = append(blankStatusItemIDs, blankStatusItemID)
-			}
-			allIssues = append(allIssues, issue)
-			if limit > 0 && keepIssue(issue) {
-				issues := keptProjectIssues(allIssues, keepIssue, limit)
-				if len(issues) >= limit {
-					c.defaultBlankProjectItemStatuses(ctx, blankStatusItemIDs)
-					if err := c.resolveBlockedByProjectState(ctx, issues); err != nil {
-						return nil, err
-					}
-					return issues, nil
-				}
-			}
-		}
-
-		if !response.Node.Items.PageInfo.HasNextPage {
-			c.defaultBlankProjectItemStatuses(ctx, blankStatusItemIDs)
-			if err := c.resolveBlockedByProjectState(ctx, allIssues); err != nil {
-				return nil, err
-			}
-			return keptProjectIssues(allIssues, keepIssue, limit), nil
-		}
-		cursor := strings.TrimSpace(response.Node.Items.PageInfo.EndCursor)
-		if cursor == "" {
-			return nil, ErrInvalidResponse
-		}
-		after = &cursor
-	}
-}
-
-func projectItemsQueryForType(queryType string) string {
-	if queryType == graphQLQueryObservedStatus {
-		return observedStatusProjectItemsQuery
-	}
-	return projectItemsQuery
-}
-
-func keptProjectIssues(allIssues []connector.Issue, keepIssue func(connector.Issue) bool, limit int) []connector.Issue {
-	issues := make([]connector.Issue, 0, len(allIssues))
-	for _, issue := range allIssues {
-		if !keepIssue(issue) {
-			continue
-		}
-		issues = append(issues, issue)
-		if limit > 0 && len(issues) >= limit {
-			return issues
-		}
-	}
-	return issues
-}
-
-func (c *Connector) normalizeProjectItem(item projectItemNode) (connector.Issue, bool, string, error) {
-	if item.Content == nil || item.Content.TypeName != "Issue" {
-		return connector.Issue{}, false, "", nil
-	}
-	c.cacheIssueRef(*item.Content)
-	statusName, statusUpdatedAt, blankStatusItemID, err := c.projectItemStatusOrDefault(item)
-	if err != nil {
-		return connector.Issue{}, false, "", err
-	}
-	return c.buildIssue(
-		*item.Content,
-		statusName,
-		singleSelectName(item.PriorityValue),
-		statusUpdatedAt,
-		projectFieldValues(item.FieldValues),
-	), true, blankStatusItemID, nil
-}
-
-func (c *Connector) projectItemStatusOrDefault(item projectItemNode) (string, *time.Time, string, error) {
-	statusName := singleSelectName(item.StatusValue)
-	if statusName != "" {
-		return statusName, singleSelectUpdatedAt(item.StatusValue), "", nil
-	}
-
-	itemID := strings.TrimSpace(item.ID)
-	if itemID == "" {
-		return "", nil, "", ErrInvalidResponse
-	}
-
-	statusName = c.detentToGitHubState(defaultProjectItemStatusState)
-	if statusName == "" {
-		return "", nil, "", nil
-	}
-	return statusName, nil, itemID, nil
-}
-
-func (c *Connector) defaultBlankProjectItemStatuses(ctx context.Context, itemIDs []string) {
-	itemIDs = uniqueNonBlank(itemIDs)
-	if len(itemIDs) == 0 {
-		return
-	}
-
-	statusName := c.detentToGitHubState(defaultProjectItemStatusState)
-	if statusName == "" {
-		return
-	}
-
-	go c.defaultBlankProjectItemStatusesAsync(ctx, itemIDs, statusName)
-}
-
-func (c *Connector) defaultBlankProjectItemStatusesAsync(parentCtx context.Context, itemIDs []string, statusName string) {
-	baseCtx := context.Background()
-	if parentCtx != nil {
-		// Default status repair is a must-finish write: detach cancellation
-		// while keeping caller values for logs and traces.
-		baseCtx = context.WithoutCancel(parentCtx)
-	}
-	ctx, cancel := context.WithTimeout(baseCtx, defaultProjectItemStatusWriteTimeout)
-	defer cancel()
-
-	if err := c.writeDefaultProjectItemStatuses(ctx, itemIDs, statusName); err != nil {
-		c.client.logger.WarnContext(ctx, "default github project item statuses failed", "count", len(itemIDs), "error", err)
-	}
-}
-
-func (c *Connector) writeDefaultProjectItemStatuses(ctx context.Context, itemIDs []string, statusName string) error {
-	itemIDs = uniqueNonBlank(itemIDs)
-	if len(itemIDs) == 0 {
-		return nil
-	}
-
-	workerCount := min(defaultProjectItemStatusWriteParallelism, len(itemIDs))
-	jobs := make(chan string)
-	errs := make(chan error, len(itemIDs))
-	var wg sync.WaitGroup
-	wg.Add(workerCount)
-	for range workerCount {
-		go func() {
-			defer wg.Done()
-			for itemID := range jobs {
-				if err := c.setProjectItemStatus(ctx, itemID, statusName); err != nil {
-					errs <- fmt.Errorf("%s: %w", itemID, err)
-				}
-			}
-		}()
-	}
-
-	for _, itemID := range itemIDs {
-		select {
-		case <-ctx.Done():
-			close(jobs)
-			wg.Wait()
-			close(errs)
-			return errors.Join(ctx.Err(), joinErrors(errs))
-		case jobs <- itemID:
-		}
-	}
-
-	close(jobs)
-	wg.Wait()
-	close(errs)
-	return joinErrors(errs)
-}
-
-func joinErrors(errs <-chan error) error {
-	var joined error
-	for err := range errs {
-		joined = errors.Join(joined, err)
-	}
-	return joined
-}
-
 func (c *Connector) resolveBlockedByProjectState(ctx context.Context, issues []connector.Issue) error {
 	byIdentifier := make(map[string]connector.Issue, len(issues))
 	for _, issue := range issues {
@@ -2169,660 +1331,6 @@ func (c *Connector) logBlockedByHydrationError(ctx context.Context, identifier s
 		return
 	}
 	c.logger.DebugContext(ctx, "github blocked-by hydration skipped", "identifier", identifier, "error", err)
-}
-
-func (c *Connector) resolveIssueProjectFields(ctx context.Context, issueID string, items *projectItemsConnection) (string, string, *time.Time, map[string]string, bool, error) {
-	if stateName, priorityName, statusUpdatedAt, fields, ok := c.projectFields(issueID, items); ok {
-		return stateName, priorityName, statusUpdatedAt, fields, true, nil
-	}
-	if items == nil || !items.PageInfo.HasNextPage {
-		return "", "", nil, nil, false, nil
-	}
-	cursor := strings.TrimSpace(items.PageInfo.EndCursor)
-	if cursor == "" {
-		return "", "", nil, nil, false, ErrInvalidResponse
-	}
-	return c.fetchProjectFieldsPage(ctx, issueID, &cursor)
-}
-
-func (c *Connector) projectFields(issueID string, items *projectItemsConnection) (string, string, *time.Time, map[string]string, bool) {
-	if items == nil {
-		return "", "", nil, nil, false
-	}
-	for _, item := range items.Nodes {
-		if item.Project != nil && item.Project.ID == c.projectID {
-			c.projectCache.SetItemID(c.projectID, issueID, item.ID)
-			return singleSelectName(item.StatusValue), singleSelectName(item.PriorityValue), singleSelectUpdatedAt(item.StatusValue), projectFieldValues(item.FieldValues), true
-		}
-	}
-	return "", "", nil, nil, false
-}
-
-func (c *Connector) fetchProjectFieldsPage(ctx context.Context, issueID string, after *string) (string, string, *time.Time, map[string]string, bool, error) {
-	var response struct {
-		Node *struct {
-			ProjectItems projectItemsConnection `json:"projectItems"`
-		} `json:"node"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryProjectItem, projectItemForIssueQuery, map[string]any{
-		"issueId":           issueID,
-		"projectItemsFirst": projectItemsPerIssue,
-		"after":             after,
-	}, &response); err != nil {
-		return "", "", nil, nil, false, fmt.Errorf("fetch github project item fields: %w", err)
-	}
-	if response.Node == nil {
-		return "", "", nil, nil, false, ErrProjectItemNotFound
-	}
-	if stateName, priorityName, statusUpdatedAt, fields, ok := c.projectFields(issueID, &response.Node.ProjectItems); ok {
-		return stateName, priorityName, statusUpdatedAt, fields, true, nil
-	}
-	if !response.Node.ProjectItems.PageInfo.HasNextPage {
-		return "", "", nil, nil, false, nil
-	}
-	cursor := strings.TrimSpace(response.Node.ProjectItems.PageInfo.EndCursor)
-	if cursor == "" {
-		return "", "", nil, nil, false, ErrInvalidResponse
-	}
-	return c.fetchProjectFieldsPage(ctx, issueID, &cursor)
-}
-
-func (c *Connector) resolveStatusOption(ctx context.Context, githubState string) (string, string, error) {
-	metadata, err := c.resolveStatusMetadata(ctx)
-	if err != nil {
-		return "", "", err
-	}
-
-	optionID, ok := metadata.OptionIDsByName[githubState]
-	if !ok || strings.TrimSpace(optionID) == "" {
-		return "", "", fmt.Errorf("%w: %s", ErrStatusOptionNotFound, githubState)
-	}
-	return metadata.FieldID, optionID, nil
-}
-
-func (c *Connector) addAssignee(ctx context.Context, ref issueRef, login string) error {
-	login = strings.TrimSpace(login)
-	if login == "" {
-		return ErrAssigneeNotFound
-	}
-
-	var response restIssue
-	if err := c.client.REST(ctx, http.MethodPost, restIssueAssigneesPath(ref), map[string]any{
-		"assignees": []string{login},
-	}, &response); err != nil {
-		return fmt.Errorf("set github assignee: %w", err)
-	}
-	if strings.TrimSpace(response.NodeID) == "" {
-		return ErrAssigneeUpdateFailed
-	}
-	return nil
-}
-
-func (c *Connector) fetchIssueAssignees(ctx context.Context, ref issueRef) ([]assignee, error) {
-	issue, err := c.fetchRESTIssue(ctx, ref)
-	if err != nil {
-		return nil, fmt.Errorf("fetch github issue assignees: %w", err)
-	}
-	if strings.TrimSpace(issue.ID) == "" {
-		return nil, ErrAssigneeNotFound
-	}
-	return issue.Assignees.Nodes, nil
-}
-
-func (c *Connector) removeAssignees(ctx context.Context, ref issueRef, logins []string) error {
-	logins = uniqueNonBlank(logins)
-	if len(logins) == 0 {
-		return ErrAssigneeUpdateFailed
-	}
-
-	var response restIssue
-	if err := c.client.REST(ctx, http.MethodDelete, restIssueAssigneesPath(ref), map[string]any{
-		"assignees": logins,
-	}, &response); err != nil {
-		return fmt.Errorf("replace github assignee: %w", err)
-	}
-	if strings.TrimSpace(response.NodeID) == "" {
-		return ErrAssigneeUpdateFailed
-	}
-	return nil
-}
-
-func (c *Connector) resolveSingleSelectFieldOptionFromField(
-	ctx context.Context,
-	fieldName string,
-	value string,
-	field projectOptionsFieldResponse,
-) (string, string, error) {
-	fieldName = strings.TrimSpace(fieldName)
-	value = strings.TrimSpace(value)
-	if fieldName == "" || value == "" {
-		return "", "", ErrProjectFieldOptionNotFound
-	}
-
-	decoded, err := decodeProjectSingleSelectField(fieldName, &field)
-	if err != nil {
-		return "", "", err
-	}
-	if optionID := singleSelectOptionID(decoded.Options, value); optionID != "" {
-		return decoded.ID, optionID, nil
-	}
-
-	for range 3 {
-		refetched, err := c.fetchProjectField(ctx, fieldName)
-		if err != nil {
-			return "", "", err
-		}
-		decoded, err = decodeProjectSingleSelectField(fieldName, &refetched)
-		if err != nil {
-			return "", "", err
-		}
-		if optionID := singleSelectOptionID(decoded.Options, value); optionID != "" {
-			return decoded.ID, optionID, nil
-		}
-		options := singleSelectOptionsWithRequiredAtEnd(decoded.Options, []projectSingleSelectOption{ownershipOption(value)})
-		updatedOptions, err := c.updateProjectFieldOptions(ctx, decoded.ID, options)
-		if err != nil {
-			return "", "", fmt.Errorf("ensure github project field options: %w", err)
-		}
-		if optionID := singleSelectOptionID(updatedOptions, value); optionID != "" {
-			return decoded.ID, optionID, nil
-		}
-	}
-	return "", "", fmt.Errorf("%w: %s=%s", ErrProjectFieldOptionNotFound, fieldName, value)
-}
-
-func (c *Connector) fetchProjectField(ctx context.Context, fieldName string) (projectOptionsFieldResponse, error) {
-	var response struct {
-		Node *struct {
-			TypeName string                       `json:"__typename"`
-			Field    *projectOptionsFieldResponse `json:"field"`
-		} `json:"node"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryProjectMetadata, singleSelectFieldQuery, map[string]any{
-		"projectId": c.projectID,
-		"fieldName": strings.TrimSpace(fieldName),
-	}, &response); err != nil {
-		return projectOptionsFieldResponse{}, fmt.Errorf("fetch github project field: %w", err)
-	}
-	if response.Node == nil || response.Node.TypeName != "ProjectV2" {
-		return projectOptionsFieldResponse{}, ErrProjectNotFound
-	}
-	if response.Node.Field == nil {
-		return projectOptionsFieldResponse{}, fmt.Errorf("%w: %s", ErrProjectFieldNotFound, strings.TrimSpace(fieldName))
-	}
-	return *response.Node.Field, nil
-}
-
-func projectTextField(field projectOptionsFieldResponse) bool {
-	return field.TypeName == "ProjectV2Field" &&
-		strings.EqualFold(strings.TrimSpace(field.DataType), "TEXT") &&
-		strings.TrimSpace(field.ID) != ""
-}
-
-func (c *Connector) resolveStatusMetadata(ctx context.Context) (statusMetadata, error) {
-	if metadata, ok := c.statusCache.Get(c.projectID); ok {
-		return metadata, nil
-	}
-
-	var response struct {
-		Node *struct {
-			Field *struct {
-				ID      string `json:"id"`
-				Options []struct {
-					ID   string `json:"id"`
-					Name string `json:"name"`
-				} `json:"options"`
-			} `json:"field"`
-		} `json:"node"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryProjectMetadata, statusFieldQuery, map[string]any{"projectId": c.projectID}, &response); err != nil {
-		return statusMetadata{}, fmt.Errorf("fetch github status field: %w", err)
-	}
-	if response.Node == nil || response.Node.Field == nil || strings.TrimSpace(response.Node.Field.ID) == "" {
-		return statusMetadata{}, ErrStatusFieldNotFound
-	}
-
-	metadata := statusMetadata{
-		FieldID:         strings.TrimSpace(response.Node.Field.ID),
-		OptionIDsByName: make(map[string]string, len(response.Node.Field.Options)),
-	}
-	for _, option := range response.Node.Field.Options {
-		name := strings.TrimSpace(option.Name)
-		id := strings.TrimSpace(option.ID)
-		if name == "" || id == "" {
-			continue
-		}
-		metadata.OptionIDsByName[name] = id
-	}
-	c.statusCache.Set(c.projectID, metadata)
-	return metadata, nil
-}
-
-type projectItemStatus struct {
-	ID         string
-	StatusName string
-}
-
-func (c *Connector) resolveProjectItem(ctx context.Context, issueID string) (projectItemStatus, error) {
-	return c.fetchProjectItemPage(ctx, issueID, nil)
-}
-
-func (c *Connector) fetchProjectItemPage(ctx context.Context, issueID string, after *string) (projectItemStatus, error) {
-	var response struct {
-		Node *struct {
-			ProjectItems projectItemsConnection `json:"projectItems"`
-		} `json:"node"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryProjectItem, projectItemForIssueQuery, map[string]any{
-		"issueId":           issueID,
-		"projectItemsFirst": projectItemsPerIssue,
-		"after":             after,
-	}, &response); err != nil {
-		return projectItemStatus{}, fmt.Errorf("fetch github project item: %w", err)
-	}
-	if response.Node == nil {
-		return projectItemStatus{}, ErrProjectItemNotFound
-	}
-
-	for _, item := range response.Node.ProjectItems.Nodes {
-		if item.Project != nil && item.Project.ID == c.projectID && strings.TrimSpace(item.ID) != "" {
-			c.projectCache.SetItemID(c.projectID, issueID, item.ID)
-			return projectItemStatus{
-				ID:         item.ID,
-				StatusName: singleSelectName(item.StatusValue),
-			}, nil
-		}
-	}
-	if !response.Node.ProjectItems.PageInfo.HasNextPage {
-		return projectItemStatus{}, ErrProjectItemNotFound
-	}
-	cursor := strings.TrimSpace(response.Node.ProjectItems.PageInfo.EndCursor)
-	if cursor == "" {
-		return projectItemStatus{}, ErrProjectItemNotFound
-	}
-	return c.fetchProjectItemPage(ctx, issueID, &cursor)
-}
-
-func (c *Connector) terminalStatusUpdateBlocked(currentStatus string, targetState string) bool {
-	currentState := c.githubToDetentState(currentStatus)
-	if !stateInList(currentState, c.terminalStates) {
-		return false
-	}
-	return !stateInList(targetState, c.terminalStates)
-}
-
-func (c *Connector) stateUpdateBlockedError(issueID string, currentStatus string, targetState string) error {
-	currentState := c.githubToDetentState(currentStatus)
-	if strings.TrimSpace(currentState) == "" {
-		currentState = strings.TrimSpace(currentStatus)
-	}
-	targetDetentState := c.githubToDetentState(targetState)
-	if strings.TrimSpace(targetDetentState) == "" {
-		targetDetentState = strings.TrimSpace(targetState)
-	}
-	return &connector.StateUpdateBlockedError{
-		IssueID:      strings.TrimSpace(issueID),
-		CurrentState: strings.TrimSpace(currentState),
-		TargetState:  strings.TrimSpace(targetDetentState),
-	}
-}
-
-func (c *Connector) updateStatusFieldValue(ctx context.Context, itemID string, fieldID string, optionID string) error {
-	return c.updateProjectV2SingleSelectFieldValue(ctx, itemID, fieldID, optionID, ErrStatusUpdateFailed)
-}
-
-func (c *Connector) updateProjectV2SingleSelectFieldValue(
-	ctx context.Context,
-	itemID string,
-	fieldID string,
-	optionID string,
-	emptyResponseError error,
-) error {
-	var response struct {
-		UpdateProjectV2ItemFieldValue *struct {
-			ProjectV2Item *struct {
-				ID string `json:"id"`
-			} `json:"projectV2Item"`
-		} `json:"updateProjectV2ItemFieldValue"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryUpdateField, updateSingleSelectFieldValueMutation, map[string]any{
-		"projectId": c.projectID,
-		"itemId":    itemID,
-		"fieldId":   fieldID,
-		"optionId":  optionID,
-	}, &response); err != nil {
-		return fmt.Errorf("update github project field: %w", err)
-	}
-	if response.UpdateProjectV2ItemFieldValue == nil ||
-		response.UpdateProjectV2ItemFieldValue.ProjectV2Item == nil ||
-		strings.TrimSpace(response.UpdateProjectV2ItemFieldValue.ProjectV2Item.ID) == "" {
-		return emptyResponseError
-	}
-	return nil
-}
-
-func (c *Connector) updateProjectV2TextFieldValue(
-	ctx context.Context,
-	itemID string,
-	fieldID string,
-	text string,
-	emptyResponseError error,
-) error {
-	var response struct {
-		UpdateProjectV2ItemFieldValue *struct {
-			ProjectV2Item *struct {
-				ID string `json:"id"`
-			} `json:"projectV2Item"`
-		} `json:"updateProjectV2ItemFieldValue"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryUpdateField, updateTextFieldValueMutation, map[string]any{
-		"projectId": c.projectID,
-		"itemId":    itemID,
-		"fieldId":   fieldID,
-		"text":      text,
-	}, &response); err != nil {
-		return fmt.Errorf("update github project field: %w", err)
-	}
-	if response.UpdateProjectV2ItemFieldValue == nil ||
-		response.UpdateProjectV2ItemFieldValue.ProjectV2Item == nil ||
-		strings.TrimSpace(response.UpdateProjectV2ItemFieldValue.ProjectV2Item.ID) == "" {
-		return emptyResponseError
-	}
-	return nil
-}
-
-func (c *Connector) deleteProjectItem(ctx context.Context, itemID string) error {
-	itemID = strings.TrimSpace(itemID)
-	if itemID == "" {
-		return ErrProjectItemRemoveFailed
-	}
-
-	var response struct {
-		DeleteProjectV2Item *struct {
-			DeletedItemID string `json:"deletedItemId"`
-		} `json:"deleteProjectV2Item"`
-	}
-	if err := c.client.GraphQLWithType(ctx, graphQLQueryRemoveItem, deleteProjectItemMutation, map[string]any{
-		"projectId": c.projectID,
-		"itemId":    itemID,
-	}, &response); err != nil {
-		return fmt.Errorf("remove github project item: %w", err)
-	}
-	if response.DeleteProjectV2Item == nil || strings.TrimSpace(response.DeleteProjectV2Item.DeletedItemID) == "" {
-		return ErrProjectItemRemoveFailed
-	}
-	return nil
-}
-
-func (c *Connector) detentToGitHubState(stateName string) string {
-	stateName = strings.TrimSpace(stateName)
-	if mapped, ok := c.stateMap[stateName]; ok {
-		return strings.TrimSpace(mapped)
-	}
-	normalized := normalizeStateName(stateName)
-	for detentState, mapped := range c.stateMap {
-		if normalizeStateName(detentState) == normalized {
-			return strings.TrimSpace(mapped)
-		}
-	}
-	return stateName
-}
-
-func (c *Connector) detentToGitHubStates(stateNames []string) []string {
-	states := make([]string, 0, len(stateNames))
-	seen := make(map[string]struct{}, len(stateNames))
-	for _, stateName := range stateNames {
-		state := c.detentToGitHubState(stateName)
-		key := normalizeStateName(state)
-		if key == "" {
-			continue
-		}
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		states = append(states, state)
-	}
-	return states
-}
-
-func (c *Connector) projectStatusQuery(stateNames []string) string {
-	states := c.detentToGitHubStates(stateNames)
-	if len(states) == 0 {
-		return ""
-	}
-
-	values := make([]string, 0, len(states))
-	for _, state := range states {
-		values = append(values, projectFilterValue(state))
-	}
-	return "status:" + strings.Join(values, ",")
-}
-
-func projectFilterValue(value string) string {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return ""
-	}
-	for _, r := range value {
-		if r >= 'a' && r <= 'z' {
-			continue
-		}
-		if r >= 'A' && r <= 'Z' {
-			continue
-		}
-		if r >= '0' && r <= '9' {
-			continue
-		}
-		if r == '_' || r == '-' || r == '.' {
-			continue
-		}
-		return strconv.Quote(value)
-	}
-	return value
-}
-
-func (c *Connector) githubToDetentState(githubState string) string {
-	githubState = strings.TrimSpace(githubState)
-	if githubState == "" {
-		return ""
-	}
-	if state := c.configuredDetentState(githubState); state != "" {
-		return state
-	}
-	for detentState, mapped := range c.stateMap {
-		if normalizeStateName(mapped) == normalizeStateName(githubState) {
-			return strings.TrimSpace(detentState)
-		}
-	}
-	return githubState
-}
-
-func (c *Connector) configuredDetentState(stateName string) string {
-	stateName = normalizeStateName(stateName)
-	if stateName == "" {
-		return ""
-	}
-	for _, candidate := range c.configuredStatusStates() {
-		candidate = strings.TrimSpace(candidate)
-		if normalizeStateName(candidate) == stateName {
-			return candidate
-		}
-	}
-	return ""
-}
-
-func (c *Connector) githubIssueStateToDetentState(state string) string {
-	switch strings.ToUpper(strings.TrimSpace(state)) {
-	case "CLOSED":
-		return c.closedIssueState()
-	case "OPEN":
-		return "Open"
-	default:
-		return ""
-	}
-}
-
-func githubIssueClosed(state string) bool {
-	return strings.EqualFold(strings.TrimSpace(state), "CLOSED")
-}
-
-func (c *Connector) closedIssueState() string {
-	for _, state := range c.terminalStates {
-		if normalizeStateName(state) == "done" {
-			return state
-		}
-	}
-	for _, state := range c.terminalStates {
-		if normalizeStateName(state) == "closed" {
-			return state
-		}
-	}
-	if len(c.terminalStates) > 0 {
-		return c.terminalStates[0]
-	}
-	return "Closed"
-}
-
-func (c *Connector) priorityRank(name string) *int {
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return nil
-	}
-	rank, ok := c.priorityMap[name]
-	if !ok || rank == nil {
-		return nil
-	}
-	value := *rank
-	return &value
-}
-
-func singleSelectName(value *singleSelectValue) string {
-	if value == nil {
-		return ""
-	}
-	return strings.TrimSpace(value.Name)
-}
-
-func singleSelectOptionID(options []projectSingleSelectOption, name string) string {
-	name = strings.TrimSpace(name)
-	for _, option := range options {
-		if strings.TrimSpace(option.Name) == name {
-			return strings.TrimSpace(option.ID)
-		}
-	}
-	return ""
-}
-
-func ownershipOption(name string) projectSingleSelectOption {
-	return projectSingleSelectOption{
-		Name:        strings.TrimSpace(name),
-		Color:       "BLUE",
-		Description: "Detent ownership identity.",
-	}
-}
-
-func singleSelectOptionsWithRequiredAtEnd(current []projectSingleSelectOption, required []projectSingleSelectOption) []projectSingleSelectOption {
-	options := normalizedSingleSelectOptions(current)
-	seen := make(map[string]struct{}, len(options)+len(required))
-	for _, option := range options {
-		seen[option.Name] = struct{}{}
-	}
-	for _, option := range required {
-		input := singleSelectOptionInput(option)
-		if input.Name == "" {
-			continue
-		}
-		if _, ok := seen[input.Name]; ok {
-			continue
-		}
-		seen[input.Name] = struct{}{}
-		options = append(options, input)
-	}
-	return options
-}
-
-func singleSelectUpdatedAt(value *singleSelectValue) *time.Time {
-	if value == nil {
-		return nil
-	}
-	return parseGitHubTime(value.UpdatedAt)
-}
-
-func actorLogin(actor *actor) string {
-	if actor == nil {
-		return ""
-	}
-	return strings.TrimSpace(actor.Login)
-}
-
-func firstAssigneeLogin(assignees nodeConnection[assignee]) string {
-	logins := allAssigneeLogins(assignees)
-	if len(logins) == 0 {
-		return ""
-	}
-	return logins[0]
-}
-
-func allAssigneeLogins(assignees nodeConnection[assignee]) []string {
-	logins := make([]string, 0, len(assignees.Nodes))
-	for _, assignee := range assignees.Nodes {
-		login := strings.TrimSpace(assignee.Login)
-		if login != "" {
-			logins = append(logins, login)
-		}
-	}
-	return logins
-}
-
-func assigneeLoginReplacement(current []assignee, targetLogin string) ([]string, bool) {
-	targetLogin = strings.TrimSpace(targetLogin)
-	removeLogins := make([]string, 0, len(current))
-	alreadyAssigned := false
-	for _, candidate := range current {
-		login := strings.TrimSpace(candidate.Login)
-		if login == "" {
-			continue
-		}
-		if strings.EqualFold(login, targetLogin) {
-			alreadyAssigned = true
-			continue
-		}
-		removeLogins = append(removeLogins, login)
-	}
-	return removeLogins, alreadyAssigned
-}
-
-func projectFieldValues(values nodeConnection[projectFieldValue]) map[string]string {
-	fields := make(map[string]string, len(values.Nodes))
-	for _, value := range values.Nodes {
-		fieldName := strings.TrimSpace(value.Field.Name)
-		if fieldName == "" {
-			continue
-		}
-		fieldValue, ok := projectFieldValueString(value)
-		if !ok {
-			continue
-		}
-		fields[fieldName] = fieldValue
-	}
-	return fields
-}
-
-func projectFieldValueString(value projectFieldValue) (string, bool) {
-	switch value.TypeName {
-	case "ProjectV2ItemFieldSingleSelectValue":
-		return strings.TrimSpace(value.Name), true
-	case "ProjectV2ItemFieldTextValue":
-		return value.Text, true
-	case "ProjectV2ItemFieldNumberValue":
-		if value.Number == nil {
-			return "", false
-		}
-		return strconv.FormatFloat(*value.Number, 'f', -1, 64), true
-	default:
-		return "", false
-	}
 }
 
 type pullRequestReference struct {
@@ -3619,122 +2127,4 @@ func pullRequestCheckNames(checks []connector.PullRequestCheck) []string {
 		names = append(names, check.Name)
 	}
 	return uniqueNonBlank(names)
-}
-
-func normalizedStateSet(states []string) map[string]struct{} {
-	out := make(map[string]struct{}, len(states))
-	for _, state := range states {
-		state = normalizeStateName(state)
-		if state != "" {
-			out[state] = struct{}{}
-		}
-	}
-	return out
-}
-
-func stateSetContains(states map[string]struct{}, state string) bool {
-	_, ok := states[normalizeStateName(state)]
-	return ok
-}
-
-func stateListWithout(states []string, excluded string) []string {
-	excluded = normalizeStateName(excluded)
-	if excluded == "" {
-		return normalizeStateList(states, nil)
-	}
-	out := make([]string, 0, len(states))
-	seen := map[string]struct{}{}
-	for _, state := range states {
-		state = strings.TrimSpace(state)
-		key := normalizeStateName(state)
-		if key == "" || key == excluded {
-			continue
-		}
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		out = append(out, state)
-	}
-	return out
-}
-
-func stateInList(state string, states []string) bool {
-	normalized := normalizeStateName(state)
-	if normalized == "" {
-		return false
-	}
-	for _, candidate := range states {
-		if normalized == normalizeStateName(candidate) {
-			return true
-		}
-	}
-	return false
-}
-
-func normalizeStateList(states []string, defaults []string) []string {
-	if len(states) == 0 {
-		states = defaults
-	}
-	out := make([]string, 0, len(states))
-	seen := map[string]struct{}{}
-	for _, state := range states {
-		state = strings.TrimSpace(state)
-		if state == "" {
-			continue
-		}
-		key := normalizeStateName(state)
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		out = append(out, state)
-	}
-	return out
-}
-
-func normalizeStateName(state string) string {
-	return strings.ToLower(strings.TrimSpace(state))
-}
-
-func cloneStateMap(values map[string]string) map[string]string {
-	cloned := make(map[string]string, len(values))
-	for key, value := range values {
-		key = strings.TrimSpace(key)
-		value = strings.TrimSpace(value)
-		if key != "" && value != "" {
-			cloned[key] = value
-		}
-	}
-	return cloned
-}
-
-func clonePriorityMapWithDefault(values map[string]*int) map[string]*int {
-	if values == nil {
-		values = defaultPriorityMap()
-	}
-	cloned := make(map[string]*int, len(values))
-	for key, value := range values {
-		key = strings.TrimSpace(key)
-		if key == "" {
-			continue
-		}
-		if value == nil {
-			cloned[key] = nil
-			continue
-		}
-		rank := *value
-		cloned[key] = &rank
-	}
-	return cloned
-}
-
-func defaultPriorityMap() map[string]*int {
-	return map[string]*int{
-		"Urgent":      new(1),
-		"High":        new(2),
-		"Medium":      new(3),
-		"Low":         new(4),
-		"No priority": nil,
-	}
 }
